requirements.txt
streamlit==1.33.0
google-generativeai==0.4.1
python-dotenv==1.0.1
PyMuPDF==1.23.9
faiss-cpu==1.7.4
sentence-transformers==2.6.1

